+++ index.ts
@@ -4,4 +4,5 @@ export * from './src/axios-interceptors';
 export * from './src/plugin';
 export * from './src/module';
 export * from './src/options';
+export * from './src/middleware';
 // export * from './types';