+++ src/storage/local.ts
@@ -2,11 +2,16 @@ import {AuthStorage} from '../../types/index';
 
 export class LocalStorage implements AuthStorage {
   set(key: string, value: any) {
-    localStorage.setItem(key, value);
+    localStorage.setItem(key, JSON.stringify(value));
   }
 
   get(key: string, defaultValue: any) {
-    return localStorage.getItem(key) || defaultValue;
+    try {
+      const value = localStorage.getItem(key);
+      return JSON.parse(value!);
+    } catch {
+      return defaultValue;
+    }
   }
 
   remove(key: string) {