+++ types/index.d.ts
@@ -2,3 +2,5 @@ export * from './options';
 export * from './storage';
 export * from './middleware';
 export * from './plugin';
+export * from './vue';
+export * from './module';