+++ types/options.d.ts
@@ -1,6 +1,7 @@
 import {AxiosRequestConfig} from 'axios';
+import {CookieAttributes} from 'js-cookie';
 
-export type SupportedAuthStorage = 'local' | 'secureLs';
+export type SupportedAuthStorage = 'local' | 'secureLs' | 'cookie';
 
 export type AuthOptions = {
   endpoints: {
@@ -37,4 +38,5 @@ export type AuthOptions = {
   storage: {
     driver: SupportedAuthStorage;
   };
+  cookie?: CookieAttributes;
 };