+++ types/storage.d.ts
@@ -1,6 +1,9 @@
+import {AuthOptions} from './options.d';
 export interface AuthStorage {
   set(key: string, value: any): void;
   get<T>(key: string, defaultValue?: any): T;
   remove(key: string): void;
-  clear(): void;
+  clear(options?: AuthOptions): void;
 }
+
+export const useStorage: (driver: SupportedAuthStorage) => AuthStorage;