+++ CHANGELOG.md
@@ -1,31 +1,84 @@
-# Change Log
+# Changelog
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-
-
-**Note:** Version bump only for package @gravitano/vue-auth
-
-
-
-
-
-## 0.0.1 (2021-09-02)
+### [0.0.5](https://ssh.dev.azure.com///compare/v0.0.4...v0.0.5) (2021-10-31)
 
 
 ### Bug Fixes
 
-* **auth:** types ([c05ad13](https://git.gits.id/frontend/starter/vue-3/starterkit/commits/c05ad13cd7242c6c233a5bf8f99364b17139e4ad))
+* no console ([62f627a](https://ssh.dev.azure.com///commit/62f627a1fc02151fb52f6d8dc435d45b59350493))
+
+### [0.0.4](https://github.com/pinteraktif/smart-akmil-web/compare/v0.0.3...v0.0.4) (2021-10-29)
+
+
+### Bug Fixes
+
+* missing types ([4c303b3](https://github.com/pinteraktif/smart-akmil-web/commit/4c303b33e3b57cc67d83aa57a47c0da494ebb170))
+* update lock file ([a3846b1](https://github.com/pinteraktif/smart-akmil-web/commit/a3846b12b7a23063b2ebf0e819cdab69d04fae30))
+
+### [0.0.3](https://github.com/pinteraktif/smart-akmil-web/compare/v0.0.2...v0.0.3) (2021-10-28)
 
 
 ### Features
 
-* **auth:** added auth module for vuex ([9a4d306](https://git.gits.id/frontend/starter/vue-3/starterkit/commits/9a4d3063d0b52bf38989b99bdd982c86d26e8527))
-* **auth:** typescript support and vuex auth module ([aca813f](https://git.gits.id/frontend/starter/vue-3/starterkit/commits/aca813f0964d0b2361fa5360651259b67c11dbe4))
-* update auth packages ([1d20410](https://git.gits.id/frontend/starter/vue-3/starterkit/commits/1d20410037ff6ddb2a4a45d29e57845696b16de5))
+* add custom vue properties definition ([759bf84](https://github.com/pinteraktif/smart-akmil-web/commit/759bf84c416177aa3fdb6b0ef0b6b643b2b2301a))
+* add prepublishOnly scripts ([0a497bd](https://github.com/pinteraktif/smart-akmil-web/commit/0a497bdc1027b5dec5a5497638a846ced823d943))
+* remove .npmrc file ([66a9156](https://github.com/pinteraktif/smart-akmil-web/commit/66a9156474bf30db647d5a77d5067feb3f9c88d0))
 
 
-### Reverts
+### Bug Fixes
 
-* Revert "refactor: remove create-starter package for now" ([3495137](https://git.gits.id/frontend/starter/vue-3/starterkit/commits/34951375cabd8ebd1782d8ce9a5a353326e7d3ba))
+* typing ([5ea80b3](https://github.com/pinteraktif/smart-akmil-web/commit/5ea80b3434bbaa4a3280ffbcfecd2cbb1035ee22))
+
+### [0.0.2](https://github.com/pinteraktif/smart-akmil-web/compare/v0.0.2-alpha.5...v0.0.2) (2021-10-27)
+
+
+### Features
+
+* add `standard-version` ([04bcc06](https://github.com/pinteraktif/smart-akmil-web/commit/04bcc06416c3b96b03d166becd8d4efc1d4e9baa))
+* added auth middleware ([8775609](https://github.com/pinteraktif/smart-akmil-web/commit/8775609d2ee1bb9af0a9bb4cc33bf06b83a697b2))
+* allow to set cookie options ([7ff562a](https://github.com/pinteraktif/smart-akmil-web/commit/7ff562ab182032f1cef4412b0c8041022e83d66d))
+* build and add dist ([a95dab1](https://github.com/pinteraktif/smart-akmil-web/commit/a95dab1353155ae26dabce547f643f24961c9b12))
+* new storage driver `cookie` ([ea1ee09](https://github.com/pinteraktif/smart-akmil-web/commit/ea1ee096d747cb7a0f21e74a8eb63734a5a64d1d))
+* update readme ([5518a8d](https://github.com/pinteraktif/smart-akmil-web/commit/5518a8dbd47335f6f4dff18a89d720ab18a4a167))
+
+## [0.0.2-alpha.5](https://github.com/pinteraktif/smark-akmil-cms/compare/v0.0.2-alpha.4...v0.0.2-alpha.5) (2021-10-04)
+
+
+### Bug Fixes
+
+* handle error on secureLs ([77296eb](https://github.com/pinteraktif/smark-akmil-cms/commit/77296eb0eca9970c0b958c95c5387763a5405f96))
+
+
+### Features
+
+* improved error handling ([59a6d76](https://github.com/pinteraktif/smark-akmil-cms/commit/59a6d76d902249210dc9bb5737ba71f5f1acd8dc))
+
+
+
+## [0.0.2-alpha.4](https://github.com/pinteraktif/smark-akmil-cms/compare/v0.0.2-alpha.3...v0.0.2-alpha.4) (2021-09-23)
+
+
+
+## [0.0.2-alpha.3](https://github.com/pinteraktif/smark-akmil-cms/compare/v0.0.2-alpha.2...v0.0.2-alpha.3) (2021-09-21)
+
+
+### Features
+
+* **local storage:** always stringify the local storage value ([d7acd3e](https://github.com/pinteraktif/smark-akmil-cms/commit/d7acd3e17c0ca393c0aca0ea78164272d1a8120d))
+* **types:** set default types for login payload ([a65682d](https://github.com/pinteraktif/smark-akmil-cms/commit/a65682d8cebee587194897f875877df77ae43cc8))
+
+
+
+## [0.0.2-alpha.2](https://github.com/pinteraktif/smark-akmil-cms/compare/v0.0.1...v0.0.2-alpha.2) (2021-09-15)
+
+
+### Features
+
+* added release scripts ([8d99673](https://github.com/pinteraktif/smark-akmil-cms/commit/8d9967323d5d976fe9096682f8838ff5f8bf60f8))
+* configurable auth storage via auth options `storage.driver` ([cda7f3d](https://github.com/pinteraktif/smark-akmil-cms/commit/cda7f3d7cde0ba0fd6ecd17e791d56d3b95a3cea))
+
+
+
+## 0.0.1 (2021-09-02)